+++ python/src/trade/context.rs
@@ -47,7 +47,7 @@ impl TradeContext {
         let ctx = TradeContextSync::try_new(Arc::new(config.0.clone()), {
             let callbacks = callbacks.clone();
             move |event| {
-                handle_push_event(&callbacks.lock(), event);
+                handle_push_event(&callbacks.lock(), event, None);
             }
         })
         .map_err(ErrorNewType)?;